+++ src/ops/rope.h
@@ -12,7 +12,8 @@ class Rope : public Op
 {
 public:
   Rope (GPUDevice *dev, Command *command, const int maxlen, const int dim,
-        const Tensor::DType dtype = Tensor::FP32);
+        const Tensor::DType dtype = Tensor::FP16);
+
   absl::StatusOr<Tensor> operator() (Tensor query,
                                      const size_t offset = 0) noexcept;
   absl::Status init () noexcept override;
